src/requirements.txt
flask == 1.0.2
google-cloud-storage == 1.18.0
fpdf == 1.7.2
PyPDF2 == 1.26.0
pdfminer.six